[outbound] split outbound conversation start into chat/email/phone
POST /outbound/conversations has been removed from the public API and
replaced by /outbound/conversations/{chat,email,phone}.

StartOutboundConversation, its params type, and CustomerSource (which
has no replacement in the new API) are gone. SupportPlatform moves to
its own file now that it outlives the outbound start params.
