+++ README.md
@@ -87,4 +87,52 @@ func main() {
 }
 ```
 
+### Starting an Outbound Conversation
+
+Outbound conversations have one method per channel:
+
+```go
+rsp, err := client.StartOutboundPhoneConversation(ctx, glabs.StartOutboundPhoneConversationParams{
+    CustomerID:      "user-1234",
+    ProcedureID:     "procedure-1234",
+    ToPhoneNumber:   "+14155551234",
+    FromPhoneNumber: "+14155550000",
+})
+```
+
+```go
+rsp, err := client.StartOutboundEmailConversation(ctx, glabs.StartOutboundEmailConversationParams{
+    CustomerID:      "user-1234",
+    ProcedureID:     "procedure-1234",
+    SupportPlatform: glabs.SupportPlatformZendesk,
+    CustomerSupportPlatformIdentifiers: []*glabs.CustomerSupportPlatformIdentifier{{
+        SupportPlatform: glabs.SupportPlatformZendesk,
+        Type:            glabs.CustomerSupportPlatformIdentifierTypeZendeskSupportUser,
+        Value:           "987654321",
+    }},
+})
+```
+
+`StartOutboundChatConversation` takes the same parameters as the email variant,
+minus `Subject`.
+
+#### Migrating from `StartOutboundConversation`
+
+`StartOutboundConversation` has been removed. Pick the method matching the
+channel you were passing:
+
+| Old `Channel`  | New method                        |
+| -------------- | --------------------------------- |
+| `ChannelVoice` | `StartOutboundPhoneConversation`  |
+| `ChannelEmail` | `StartOutboundEmailConversation`  |
+| `ChannelWeb`   | `StartOutboundChatConversation`   |
+
+- `CustomerSource` has no replacement, and `CustomerID` is now always your own
+  customer ID. Identifiers for third-party support platforms go in
+  `CustomerSupportPlatformIdentifiers`, keyed by platform — Zendesk requires the
+  `zendesk_support_user` type, and Salesforce the `salesforce_contact_id` type.
+- `SupportPlatform` is required for chat and email; the highest-priority
+  connected platform is no longer selected for you. Phone has no support
+  platform: the call is always placed over voice.
+
 For more examples, see the [examples](./examples) directory.
